+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### VB -> C#
 
+* No longer incorrectly assigns events in constructor for WithEvents properties
+* Event assignments created from Handles clauses now correctly appear last in the constructor [#991](https://github.com/icsharpcode/CodeConverter/issues/991)
 
 ### C# -> VB