+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * No longer incorrectly assigns events in constructor for WithEvents properties
 * Event assignments created from Handles clauses now correctly appear last in the constructor [#991](https://github.com/icsharpcode/CodeConverter/issues/991)
 * Worked around "CONVERSION ERROR: usingKeyword" bug caused by VS 17.7.0 preview 2 [#1019](https://github.com/icsharpcode/CodeConverter/issues/1019)
+* No longer inserts null checks in query expressions [#1011](https://github.com/icsharpcode/CodeConverter/issues/1011)
 
 ### C# -> VB